Extract indexing engine
commit-id:bf2373d1
